chore(go.d/sd): rename discoverers pkgs
