+++ src/go/plugin/go.d/agent/discovery/sd/discoverer/dockersd/sim_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package dockerd
+package dockersd
 
 import (
 	"context"
