+++ src/go/plugin/go.d/agent/discovery/sd/discoverer/k8ssd/service.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package kubernetes
+package k8ssd
 
 import (
 	"context"
