+++ src/go/plugin/go.d/agent/discovery/sd/pipeline/config.go
@@ -7,9 +7,9 @@ import (
 	"fmt"
 
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/dockerd"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/kubernetes"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/netlisteners"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/dockersd"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/k8ssd"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/netlistensd"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/snmpsd"
 )
 
@@ -25,11 +25,11 @@ type Config struct {
 }
 
 type DiscoveryConfig struct {
-	Discoverer   string              `yaml:"discoverer"`
-	NetListeners netlisteners.Config `yaml:"net_listeners"`
-	Docker       dockerd.Config      `yaml:"docker"`
-	K8s          []kubernetes.Config `yaml:"k8s"`
-	SNMP         snmpsd.Config       `yaml:"snmp"`
+	Discoverer   string             `yaml:"discoverer"`
+	NetListeners netlistensd.Config `yaml:"net_listeners"`
+	Docker       dockersd.Config    `yaml:"docker"`
+	K8s          []k8ssd.Config     `yaml:"k8s"`
+	SNMP         snmpsd.Config      `yaml:"snmp"`
 }
 
 type ClassifyRuleConfig struct {
