+++ src/go/plugin/go.d/agent/discovery/sd/pipeline/pipeline.go
@@ -11,9 +11,9 @@ import (
 
 	"github.com/netdata/netdata/go/plugins/logger"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/dockerd"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/kubernetes"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/netlisteners"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/dockersd"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/k8ssd"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/netlistensd"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/discoverer/snmpsd"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/discovery/sd/model"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/hostinfo"
@@ -79,7 +79,7 @@ func (p *Pipeline) registerDiscoverers(conf Config) error {
 		switch cfg.Discoverer {
 		case "net_listeners":
 			cfg.NetListeners.Source = conf.Source
-			td, err := netlisteners.NewDiscoverer(cfg.NetListeners)
+			td, err := netlistensd.NewDiscoverer(cfg.NetListeners)
 			if err != nil {
 				return fmt.Errorf("failed to create '%s' discoverer: %v", cfg.Discoverer, err)
 			}
@@ -90,14 +90,14 @@ func (p *Pipeline) registerDiscoverers(conf Config) error {
 				continue
 			}
 			cfg.Docker.Source = conf.Source
-			td, err := dockerd.NewDiscoverer(cfg.Docker)
+			td, err := dockersd.NewDiscoverer(cfg.Docker)
 			if err != nil {
 				return fmt.Errorf("failed to create '%s' discoverer: %v", cfg.Discoverer, err)
 			}
 			p.discoverers = append(p.discoverers, td)
 		case "k8s":
 			for _, k8sCfg := range cfg.K8s {
-				td, err := kubernetes.NewKubeDiscoverer(k8sCfg)
+				td, err := k8ssd.NewKubeDiscoverer(k8sCfg)
 				if err != nil {
 					return fmt.Errorf("failed to create '%s' discoverer: %v", cfg.Discoverer, err)
 				}
